Add HTTPS support with SPDY
