--- .csslintrc
@@ -1,19 +1,0 @@
-{
-  "box-model": false,
-  "universal-selector": false,
-  "regex-selectors": false,
-  "unqualified-attributes": false,
-  "box-sizing": false,
-  "unique-headings": false,
-  "qualified-headings": false,
-  "font-sizes": false,
-  "duplicate-properties": false,
-  "duplicate-background-images": false,
-  "outline-none": false,
-  "adjoining-classes": false,
-  "floats": false,
-  "text-indent": false,
-  "known-properties": false,
-  "gradients": false,
-  "font-faces": false
-}
