+++ .gitignore
@@ -4,4 +4,5 @@ node_modules
 rethinkdb_data
 *.log
 *.orig
-
+ssl/
+!ssl/README.md
