+++ gulp-includes/handleError.js
@@ -1,0 +1,6 @@
+export default function handleError (err) {
+  if (err) {
+    console.error(`Error during the gulp build: ${err}`)
+    process.exit(1)
+  }
+}
