+++ gulp-includes/styles.js
@@ -1,0 +1,40 @@
+import gulp from 'gulp'
+import postcss from 'gulp-postcss'
+import cssImport from 'postcss-import'
+import cssnano from 'cssnano'
+import cssnext from 'cssnext'
+import customMedia from 'postcss-custom-media'
+import rgbaFallback from 'postcss-color-rgba-fallback'
+import vmin from 'postcss-vmin'
+import pixrem from 'pixrem'
+import nested from 'postcss-nested'
+import mqPacker from 'css-mqpacker'
+import autoprefixer from 'autoprefixer'
+import sourcemaps from 'gulp-sourcemaps'
+import browsersync from 'browser-sync'
+
+export default gulp.task('styles', () => {
+  return gulp.src('../src/css/main.css')
+    .pipe(sourcemaps.init())
+    .pipe(postcss([
+      cssImport(),
+      nested(),
+      cssnano(),
+      cssnext(),
+      customMedia(),
+      rgbaFallback(),
+      vmin(),
+      pixrem(),
+      mqPacker({
+        sort: true
+      }),
+      autoprefixer({
+        browsers: ['last 2 versions']
+      })
+    ]))
+    .pipe(sourcemaps.write('.'))
+    .pipe(gulp.dest('../dist/css/'))
+    .pipe(browsersync.stream({
+      once: true
+    }))
+})
