+++ package.json
@@ -44,6 +44,7 @@
     "helmet": "^0.15.0",
     "history": "1.13.x",
     "hpp": "^0.2.0",
+    "http2": "^3.2.0",
     "lodash.escape": "^3.0.0",
     "lodash.unescape": "^3.0.0",
     "pm2": "^0.15.10",
@@ -54,7 +55,8 @@
     "redux": "^3.0.4",
     "rethinkdb": "^2.2.0",
     "socket.io": "^1.3.7",
-    "socket.io-client": "^1.3.7"
+    "socket.io-client": "^1.3.7",
+    "spdy": "^3.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.1.2",
