+++ server.js
@@ -1,5 +1,6 @@
+import fs from 'fs'
 import express from 'express'
-import http from 'http'
+import spdy from 'spdy'
 import socketIO from 'socket.io'
 import compression from 'compression'
 import bodyParser from 'body-parser'
@@ -13,13 +14,18 @@ import { Provider } from 'react-redux'
 import { ReduxRouter, routerStateReducer } from 'redux-router'
 import { match, reduxReactRouter } from 'redux-router/server'
 import escape from 'lodash.escape'
+import config from './config/environments'
 import bkmrkdRoutes from './src/js/main'
 import { bookmarks, endOfBookmarks, networkState, page, toaster } from './src/js/helpers/reducers'
 
-const app = express()
-const server = http.Server(app)
-const io = socketIO(server)
 const env = process.env.NODE_ENV || 'development'
+const spdyOptions = {
+  key: fs.readFileSync(config[env].certKey),
+  cert: fs.readFileSync(config[env].cert)
+}
+const app = express()
+const server = spdy.createServer(spdyOptions, app)
+const io = socketIO(server)
 
 let connection
 let bkmrkd
@@ -127,8 +133,6 @@ app.use(helmet.xssFilter())
 app.use(helmet.frameguard('deny'))
 app.use(helmet.ieNoOpen())
 app.use(helmet.noSniff())
-
-http.globalAgent.maxSockets = 1000
 
 app.route(/^\/(colophon)?$/)
   .get((req, res) => {
